+++ src/main/java/com/nl/Nutso/model/dto/BookDetailDTO.java
@@ -3,14 +3,13 @@ package com.nl.Nutso.model.dto;
 import com.nl.Nutso.model.enums.BookConditionEnum;
 import com.nl.Nutso.model.enums.CategoryEnum;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record BookDetailDTO(
         UUID uuid,
         String title,
         String author,
-        BigDecimal price,
+        double price,
         BookConditionEnum bookCondition,
         String additionalInfo,
         CategoryEnum category,
@@ -19,7 +18,7 @@ public record BookDetailDTO(
         Boolean isAvailable) {
 
 
-    public BookDetailDTO(UUID uuid, String title, String author, BigDecimal price,
+    public BookDetailDTO(UUID uuid, String title, String author, double price,
                          BookConditionEnum bookCondition, String additionalInfo,
                          CategoryEnum category, Integer year, String imageUrl,
                          Boolean isAvailable) {
